fix: access to unexisting key.
Fix the access to unexisting keys in the classification dictionary by adding a
verification of the existence of the predicted class in the dictionary. If there
is no key for the predicted class, then the system will create a new key to hold
and count the classification for the new predicted class.